Implement CLI::RewriteLine()
This will add Terminal_MoveCursorY_OneUp() and
Terminal_EraseFromCursorToEOL() to class CLI in order
to control the terminal and to overwrite a line.

RewriteLine() is implemented. Depending on flag
"allowRewriteLine" it either rewrites the previous line or
it behaves like WriteLine().

"allowRewriteLine" is cleared by an ordinary WriteLine()
and whenever the user is asked to enter text into the
terminal. The latter also ensures that the flag is cleared
when a command handler is invoked.