Live with unexpected behaviour of std::sto*
On some platforms, std::sto* accepts prefix 0b/0B
for binaray values as it does with 0x/0X for
hexadecimal values. However, prefix 0b/0B is not
documented and should not be accepted. GPCC
relies on this.

If glibc is used, GPCC will now check itself for
prefix 0b/0B if required.

The issue has first been observed on
kubuntu 24.04 with gcc 13.2.0 and glibc 2.39